usb: gadget: f_fs: Remove unused variables
With commit [1], our original intention was to just remove
ipc log string and resolve the uninitialized variable warn
for our benefits. It seems commit [1] made that variable's
usecase unused and also the ones who are dependent on it
and the similars.

This removes all of the unused variables detected by
GCC after commit [1].

[1]: usb: gadget: f_fs: Make ipc log strings no-ops

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: kawaaii <kawaaii@nocturn9x.space>