exfat: Allow using `exfat_zeroed_cluster()` on downstream
We have backported sync_blockdev_range() from mainline with
commit [1].
We can now use it without any restriction. Therefore, allow
the usage of sync_blockdev_range() for kernel revisions greater
or equal to 4.19 and remove the direct call to filemap_write_and_wait_range()
which was previously done to bypass the use of sync_blockdev_range().

[1]: https://github.com/Atom-X-Devs/android_kernel_xiaomi_scarlet/commit/ec18466532faa5459ce4778609cac44539081a77

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: kawaaii <kawaaii@nocturn9x.space>